Mordil. Cages left seems to understand that a hand is locked within, still debugging right hand. Neither attack but that won't last the weekend. Most of it is based on timers, such as breaking in/out.
Going to add a HARD reset to the hand where if it hasn't moved for the last 5 seconds or so it should go back to the start position. Not smooth but otherwise debris or detritus could get in the way.

I'll need to use Your cutscene editor I think if we want a pacifist (undertale) ending. Ropes work, use Rope prefab whereever one should spawn and it will generate until it reaches a "platform" object.

Cages exist as a prefab to allow easier management of their information.