Updated things related to the Magnetic Gloves. Players and objects will have a constant force added to them when being pushed/pulled. This force can only be added to when the player is pushing vertically against an object. Fixed an issue that allowed out of range items to be selected as the target.
Metal items are now gray instead of orange.
Metal items now turn red/blue when they are push pulled.
A similarly colored line is drawn between the player and the object.

Added methods to the UI Manager to allow for revealing of multiline item descriptions.
Item descriptions are now an array of strings instead of one.
Added Info string to item to be used in menu to allow players to check what an item can do.